Refactor query filter configuration loader
- Renamed CriteriaConfigLoader to ConfigLoader and updated import paths accordingly.
- Changed the configuration JSON file from cohort-criteria-config-2.json to atlas-config.json.
- Removed the CriteriaConfigLoader class and its associated tests.
- Updated references in AtlasConverter and loadAtlasCohortDefinition to use the new ConfigLoader.
- Adjusted tests to reflect changes in the configuration loader structure and data.
