+++ ui/apps/vue-mri-ui-lib/src/query-filter/config/atlas-config.json
