+++ ui/apps/vue-mri-ui-lib/src/query-filter/utils/AtlasConverter.ts
@@ -22,7 +22,7 @@ import {
 } from '../types/AtlasTypes'
 
 import type { ConceptSetItemDisplay, SelectedConceptSet, StoredConceptItem } from '../types/ConceptSetTypes'
-import type CriteriaConfigLoader from './CriteriaConfigLoader'
+import type ConfigLoader from './ConfigLoader'
 
 export interface ConceptSetMapping {
   name: string
@@ -111,7 +111,7 @@ const convertConceptSetArrayToAttribute = (
   attributeId: string,
   conceptArray: Concept[],
   eventType: string,
-  configLoader?: typeof CriteriaConfigLoader
+  configLoader?: typeof ConfigLoader
 ): QueryFilterAttribute => {
   const conceptItems = convertAtlasConceptsToInternal(conceptArray)
 
@@ -189,7 +189,7 @@ const convertConceptSetItemToSelected = (item: ConceptSetItemDisplay): SelectedC
 export const convertAtlasToFilters = (
   atlasJson: AtlasCohortDefinition,
   availableConceptSets: ConceptSetItemDisplay[] = [],
-  configLoader?: typeof CriteriaConfigLoader
+  configLoader?: typeof ConfigLoader
 ): QueryFilterCriteriaManager => {
   try {
     if (!atlasJson) {
